% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ratios.R
\name{ratio.append_smallest}
\alias{ratio.append_smallest}
\title{ratio.append_smallest}
\usage{
ratio.append_smallest(Ratios, vars = NULL)
}
\arguments{
\item{Ratios}{list, data.frame or data.table, which is the output after using the function \code{\link{ratio.DT1_DT2}}}

\item{vars}{optional, character vector of column names of DT1 and DT2, default is function \code{\link{select.VarsElements}}.
Please make sure the columns given in \code{vars} are of class numeric.}
}
\value{
list with [[1]] being the data set from the input with one column added containing the smallest ratio of all variables given in \code{vars}.
If the input was a list with one element named "ratios_error" the returned list contains a second element [[2]] "ratios_error" also with the appended columns.
}
\description{
The function appends for each row the smallest ratio DT1/DT2 in the column \emph{ratio_smallest}.
The name of the column which contained the smallest ratio is appended in the column \emph{ratio_smallest_Elem}.
This function is basically a sub-function for the function \code{\link{Correction.AdheringParticles}}.
}
\seealso{
Other ratio functions: \code{\link{Correction.AdheringParticles}},
  \code{\link{preparation.DT2}},
  \code{\link{ratio.DT1_DT2}}
}
\author{
Solveig Pospiech
}
