% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{ratios-package}
\alias{ratios}
\alias{ratios-package}
\title{Ratios of environmental data}
\description{
The package provides functions for calculating ratios between two data sets containing environmental data like concentration of elements.
Ratios can be calculated either by method "simple", "clr" or "alr".
Additionally for plants the amount of adhering particles on plants can be estimated and the element concentration corrected by subtraction.
}
